refactor(architecture): FSD features 레이어 구축 및 기능 단위 통합
FSD 아키텍처 리팩토링 3단계로, 기존 `components`, `hooks`, `utils` 등에 분산되어 있던 기능 관련 코드들을 `features` 레이어로 통합합니다.

'콘텐츠 관리', 'AI 미디어 요청', '실시간 알림', '인증' 등 5개의 핵심 기능을 사용자 시나리오 중심으로 재구성했습니다. 각 피처 슬라이스 내부에 UI(`ui`), 비즈니스 로직(`model`), 유틸리티(`lib`)를 함께 배치하여 기능의 응집도와 독립성을 극대화했습니다.

**주요 변경 사항:**
- `src/features` 디렉토리 생성 및 5개의 기능 슬라이스 설계
- `src/components`의 UI 컴포넌트와 `hooks`, `stores` 등의 로직을 기능 단위로 통합
- 35개 이상 파일의 `import` 경로를 새로운 `features` 경로로 업데이트
- 기존 `hooks`, `stores`, `utils`의 기능 종속적 코드들을 모두 정리하여 해당 폴더들을 비움
