+++ src/common/constants/imagePaths.js
@@ -1,0 +1,12 @@
+// 랜딩 페이지 이미지 경로 상수
+import hero_section from '../../assets/images/LandingImg/herosection.png';
+import insight from '../../assets/images/LandingImg/insight.png';
+import automation from '../../assets/images/LandingImg/storiesStaring.png';
+import optimization from '../../assets/images/LandingImg/perfectTiming.png';
+
+export const IMAGE_PATHS = {
+  hero_section,
+  insight,
+  automation,
+  optimization
+};
