+++ src/common/utils/pageConfig.js
