+++ src/common/utils/videoUrlUtils.js
