+++ src/components/dashboard/dashboard_header.jsx
@@ -6,7 +6,7 @@
 import React, { useCallback } from 'react';
 import { motion } from 'framer-motion';
 import { Sun, Moon } from 'lucide-react';
-import { usePageStore } from '../../stores/page_store.js';
+import { usePageStore } from '../../common/stores/pageStore.js';
 import { get_header_info } from '../../domain/dashboard/model/dashboardUtils.js';
 import GlassCard from '../../common/ui/glass-card.jsx';
 import Notification from '../../common/ui/notification.jsx';
