+++ src/components/landing/HeroSection.jsx
@@ -1,7 +1,7 @@
 import { motion } from 'framer-motion';
 import { Link } from 'react-router-dom';
 import { ArrowRight } from 'lucide-react';
-import { IMAGE_PATHS } from '../../utils/image_paths.js';
+import { IMAGE_PATHS } from '../../common/constants/imagePaths.js';
 import Section from '../../common/ui/Section.jsx';
 
 /**
