+++ src/components/landing/TransformSection.jsx
@@ -1,6 +1,6 @@
 import { motion } from 'framer-motion';
 import { BarChart3, Rocket, Target } from 'lucide-react';
-import { IMAGE_PATHS } from '../../utils/image_paths.js';
+import { IMAGE_PATHS } from '../../common/constants/imagePaths.js';
 import Section from '../../common/ui/Section.jsx';
 
 /**
