+++ src/features/ai-media-request/ui/AiMediaRequestModal.jsx
@@ -7,12 +7,12 @@ import React, { useEffect, useCallback } from 'react';
 import { createPortal } from 'react-dom';
 import { motion, AnimatePresence } from 'framer-motion';
 import { X } from 'lucide-react';
-import { Button } from '../../common/ui/button.jsx';
-import SuccessModal from '../../common/ui/success_modal.jsx';
-import LocationSelector from '../../common/ui/location_selector.jsx';
-import ImageUploader from '../../common/ui/image_uploader.jsx';
-import NaturalPromptInput from '../../common/ui/natural_prompt_input.jsx';
-import { useMediaRequestForm } from '../../hooks/use_media_request_form.js';
+import { Button } from '../../../common/ui/button.jsx';
+import SuccessModal from '../../../common/ui/success_modal.jsx';
+import LocationSelector from '../../../common/ui/location_selector.jsx';
+import ImageUploader from '../../../common/ui/image_uploader.jsx';
+import NaturalPromptInput from '../../../common/ui/natural_prompt_input.jsx';
+import { useMediaRequestForm } from '../model/useMediaRequestForm.js';
 
 /**
  * AI 미디어 제작 요청 모달 컴포넌트
