+++ src/features/content-modals/ui/ContentListView.jsx
@@ -1,15 +1,15 @@
 import { useState, useEffect, useRef } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import { ChevronDown, ChevronLeft, ChevronRight, Clock, Image, MessageSquare, ThumbsUp, ArrowBigUp, Eye } from 'lucide-react';
-import { Pagination, PaginationContent, PaginationEllipsis, PaginationItem } from "../../common/ui/pagination.jsx";
-import GlassCard from '../../common/ui/glass-card.jsx';
-import { getYouTubeVideosByChannelId } from '../../common/api/api.js';
-import { useYouTubeStore } from '../../domain/youtube/model/store.js';
-import { usePlatformStore } from '../../domain/platform/model/store.js';
-import { mockContentData } from '../../utils/mock-data.js';
-import { use_content_modals } from '../../hooks/use_content_modals.jsx';
-import ContentPreviewModal from './content_preview_modal.jsx';
-import { usePageStore } from '../../stores/page_store.js';
+import { Pagination, PaginationContent, PaginationEllipsis, PaginationItem } from "../../../common/ui/pagination.jsx";
+import GlassCard from '../../../common/ui/glass-card.jsx';
+import { getYouTubeVideosByChannelId } from '../../../common/api/api.js';
+import { useYouTubeStore } from '../../../domain/youtube/model/store.js';
+import { usePlatformStore } from '../../../domain/platform/model/store.js';
+import { mockContentData } from '../../../common/utils/mockData.js';
+import { use_content_modals } from '../model/useContentModals.js';
+import ContentPreviewModal from './ContentPreviewModal.jsx';
+import { usePageStore } from '../../../common/stores/pageStore.js';
 
 function ContentListView({
   selectedPlatform,
