+++ vite.config.js
@@ -6,14 +6,15 @@ export default defineConfig({
   resolve: {
     alias: {
       '@': '/src'
-    }
+    },
+    extensions: ['.js', '.jsx', '.ts', '.tsx', '.json']
   },
   server: {
     proxy: {
       '/api': {
         target: 'http://localhost:8080', // 백엔드 주소
         changeOrigin: true,              // Origin 헤더 변경
-        secure: false,                   // HTTPS 인증서 무시 
+        secure: false,                   // HTTPS 인증서 무시
       }
     }
   }
